# Conflicts:
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/app/observer/TrimObserver.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/dao/state/derived/BasicDbTable.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/dao/state/derived/DerivedDbTable.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/service/appdata/TrimService.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/service/blockchain/BlockchainProcessorImpl.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/shard/ShardEngineImpl.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/shard/ShardService.java
#	apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/shard/observer/ShardObserver.java
#	apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/app/observer/TrimObserverScheduleLogicTest.java
#	apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/app/observer/TrimObserverTest.java
#	apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/chainid/BlockchainConfigTest.java
#	apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/service/appdata/TrimServiceTest.java
#	apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/shard/ShardServiceIntegrationTest.java
#	apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/shard/ShardServiceTest.java
#	apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/shard/observer/ShardObserverIntegrationTest.java
#	apl-core/src/test/java/com/apollocurrency/aplwallet/apl/core/shard/observer/ShardObserverTest.java

+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/dao/state/derived/MinMaxValue.java
@@ -7,11 +7,13 @@ package com.apollocurrency.aplwallet.apl.core.dao.state.derived;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class MinMaxValue {
-    private long min;
-    private long max;
+    private BigDecimal min;
+    private BigDecimal max;
     private String column;
     private long count;
     private int height;
